Added date picker to expense screen and adjusted toast duration

+++ app/src/main/java/com/sofato/krone/ui/expenses/ExpenseListScreen.kt
@@ -57,7 +57,7 @@ fun ExpenseListScreen(
             val result = snackbarHostState.showSnackbar(
                 message = deletedMessage,
                 actionLabel = undoLabel,
-                duration = SnackbarDuration.Long,
+                duration = SnackbarDuration.Short,
             )
             if (result == SnackbarResult.ActionPerformed) {
                 viewModel.undoDelete()
